(PC-20623)[API] feat: Notification cultural partner when he can create collective offer
