+++ api/src/pcapi/core/offerers/api.py
@@ -25,6 +25,7 @@ import pcapi.core.finance.models as finance_models
 import pcapi.core.history.api as history_api
 import pcapi.core.history.models as history_models
 import pcapi.core.mails.transactional as transactional_mails
+from pcapi.core.mails.transactional import send_eac_offerer_activation_email
 from pcapi.core.offerers import models as offerers_models
 import pcapi.core.offers.models as offers_models
 import pcapi.core.providers.models as providers_models
@@ -613,6 +614,11 @@ def validate_offerer(offerer: models.Offerer, author_user: users_models.User) ->
                 "Could not send validation confirmation email to offerer",
                 extra={"offerer": offerer.id},
             )
+    for managed_venue in managed_venues:
+        if managed_venue.adageId:
+            emails = offerers_repository.get_emails_by_venue(managed_venue)
+            send_eac_offerer_activation_email(managed_venue, list(emails))
+            break
 
 
 def reject_offerer(
